fix : 1. util의 클래스명 변경 2. 이미 있는 예외와 이름이 중복되어 잘못 참조되는 문제 해결 3. createRepositoryId 메서드 사용 추가 (#8)
