+++ src/main/java/com/econo/hackday/contentsproxyblog/exception/GithubFileNotFoundException.java
@@ -1,8 +1,8 @@
 package com.econo.hackday.contentsproxyblog.exception;
 
-public class FileNotFoundException extends RuntimeException {
+public class GithubFileNotFoundException extends RuntimeException {
 
-    public FileNotFoundException(String path) {
+    public GithubFileNotFoundException(String path) {
         super(path + "에 파일을 찾을수  없습니다.");
     }
 }
